pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiogram>=3.24.0",
    "aiohttp>=3.13.3",
    "flask>=3.1.2",
    "gunicorn>=25.0.3",
    "psycopg2-binary>=2.9.11",
    "requests>=2.32.5",
]
